# api/docs/swagger.yaml
definitions:
...
  booking.ListSubscriptionGroupResponse:
    properties:
      subscription_group:
        items:
          $ref: '#/definitions/booking.SubscriptionGroup'
        type: array
    type: object
  booking.ListSubscriptionPersonalResponse:
    properties:
      subscription_personal:
        items:
          $ref: '#/definitions/booking.SubscriptionPersonal'
        type: array
    type: object
...
  booking.SubscriptionGroup:
    properties:
      capacity:
        type: integer
      coach_id:
        type: string
      count:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: integer
      description:
        type: string
      duration:
        type: integer
      gym_id:
        type: string
      id:
        type: string
      price:
        type: integer
      time:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  booking.SubscriptionPersonal:
    properties:
      count:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: integer
      description:
        type: string
      duration:
        type: integer
      gym_id:
        type: string
      id:
        type: string
      price:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /v1/subscription-group:
    get:
      consumes:
      - application/json
      description: Get a list of subscription group records.
      parameters:
      - description: Filter by gym ID
        in: query
        name: gym_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListSubscriptionGroupResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Subscription Group
      tags:
      - SubscriptionGroup
    post:
      consumes:
      - application/json
      description: Create a new subscription group with the provided details.
      parameters:
      - description: Subscription Group details
        in: body
        name: subscription_group
        required: true
        schema:
          $ref: '#/definitions/booking.SubscriptionGroup'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/booking.SubscriptionGroup'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Subscription Group
      tags:
      - SubscriptionGroup
  /v1/subscription-group/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a subscription group by its ID.
      parameters:
      - description: Subscription Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Subscription Group
      tags:
      - SubscriptionGroup
    get:
      consumes:
      - application/json
      description: Get details of a subscription group by its ID.
      parameters:
      - description: Subscription Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.SubscriptionGroup'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Subscription Group by ID
      tags:
      - SubscriptionGroup
    put:
      consumes:
      - application/json
      description: Update an existing subscription group.
      parameters:
      - description: Subscription Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated subscription group
        in: body
        name: subscription_group
        required: true
        schema:
          $ref: '#/definitions/booking.SubscriptionGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.SubscriptionGroup'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Subscription Group
      tags:
      - SubscriptionGroup
  /v1/subscription-personal:
    get:
      consumes:
      - application/json
      description: Get a list of subscription personal records.
      parameters:
      - description: Filter by gym ID
        in: query
        name: gym_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListSubscriptionPersonalResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Subscription Personal
      tags:
      - SubscriptionPersonal
    post:
      consumes:
      - application/json
      description: Create a new subscription personal with the provided details.
      parameters:
      - description: Subscription Personal details
        in: body
        name: subscription_personal
        required: true
        schema:
          $ref: '#/definitions/booking.SubscriptionPersonal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/booking.SubscriptionPersonal'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Subscription Personal
      tags:
      - SubscriptionPersonal
  /v1/subscription-personal/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a subscription personal by its ID.
      parameters:
      - description: Subscription Personal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Subscription Personal
      tags:
      - SubscriptionPersonal
    get:
      consumes:
      - application/json
      description: Get details of a subscription personal by its ID.
      parameters:
      - description: Subscription Personal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.SubscriptionPersonal'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Subscription Personal by ID
      tags:
      - SubscriptionPersonal
    put:
      consumes:
      - application/json
      description: Update an existing subscription personal.
      parameters:
      - description: Subscription Personal ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated subscription personal
        in: body
        name: subscription_personal
        required: true
        schema:
          $ref: '#/definitions/booking.SubscriptionPersonal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.SubscriptionPersonal'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Subscription Personal
      tags:
      - SubscriptionPersonal
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"